+++ packages/react/src/actions/sendTransaction.ts
@@ -1,6 +1,8 @@
 import { Transaction } from '@mysten/sui/transactions';
 import { Signer, SubmittableExtrinsic } from '@polkadot/api/types';
 import { VersionedTransaction as SolanaVersionedTransaction } from '@solana/web3.js';
+import { Cell } from '@ton/ton';
+import { CHAIN } from '@tonconnect/ui-react';
 import { sendTransaction as sendEVMTransaction } from '@wagmi/core';
 import { Address as EVMAddress } from 'viem';
 import { ChainData, ChainType, ConnectionOrConfig } from '../types/index.js';
@@ -33,6 +35,14 @@ type TransactionArgs<CType extends ChainType> = CType extends 'evm' | 'tron'
       : CType extends 'sui'
         ? {
             tx: Transaction;
+        }
+      : CType extends 'ton'
+        ? {
+            tonArgs: {
+              validUntil: number; // transaction deadline in unix epoch seconds.
+              network?: CHAIN; // (MAINNET: "-239" & TESTNET: "-3")
+              payload?: string;
+              stateInit?: string;
           }
         : never;
 
@@ -156,6 +166,39 @@ export const sendTransactionToChain = (async ({ chain, to, from, value, args, co
     });
 
     return result.digest;
+  }
+
+  if (chain.type === 'ton') {
+    const { tonArgs } = args as TransactionArgs<'ton'>;
+    const messages: Array<{
+      address: string;
+      amount: string;
+      payload?: string;
+      stateInit?: string;
+    }> = [
+      {
+        address: to,
+        amount: value.toString(),
+        payload: tonArgs.payload,
+        stateInit: tonArgs.stateInit,
+      },
+    ];
+    const transaction = {
+      from,
+      messages,
+      network: tonArgs.network,
+      validUntil: tonArgs.validUntil,
+    };
+
+    const walletConnector = config.connector as WalletInstance<'ton'>;
+    // send transaction to TON chain
+    const tx = await walletConnector.sendTransaction(transaction);
+
+    const cell = Cell.fromBase64(tx.boc);
+    const buffer = cell.hash();
+    const hashHex = buffer.toString('hex');
+
+    return { txHash: hashHex };
   }
 
   throw new Error('Chain not supported');