+++ packages/react/src/actions/waitForTransaction.ts
@@ -1,3 +1,4 @@
+import { Address } from '@ton/ton';
 import { waitForTransactionReceipt } from '@wagmi/core';
 import { ReplacementReturnType } from 'viem';
 import { ChainData, ChainType, ConnectionOrConfig, TransactionReceipt } from '../types/index.js';
@@ -19,7 +20,12 @@ export type WatchTransactionOverrides<C extends ChainType> = DefaultOverrides &
       ? {
           maxSupportedTransactionVersion: number;
         }
-      : any);
+      : C extends 'ton'
+        ? {
+            accountAddress: string;
+            lt: string;
+          }
+        : any);
 
 export type DefaultTransactionParams = {
   txHash: string;
@@ -170,6 +176,15 @@ export const waitForTransaction = (async ({ chain, config, overrides, transactio
             showBalanceChanges: true,
           },
         });
+      }
+
+  if (chain.type === 'ton') {
+    const _overrides = (overrides || {}) as WatchTransactionOverrides<'ton'>;
+    const { txHash } = transactionParams as TransactionParams<'ton'>;
+
+    const receipt = await pollCallback(
+      async () => {
+        return await config.tonClient.getTransaction(Address.parse(_overrides.accountAddress), _overrides.lt, txHash);
       },
       {
         interval: overrides?.interval || DEFAULT_POLLING_INTERVAL,