+++ packages/react/src/chains/index.ts
@@ -6,5 +6,6 @@ export * from './solana.js';
 export * from './solana.testnet.js';
 export * from './sui.js';
 export * from './sui.testnet.js';
+export * from './ton.js';
 export * from './tron.js';
 export * from './tron.shasta.js';