+++ packages/react/src/hooks/useConnect.ts
@@ -4,10 +4,12 @@ import { useWallet as useSolanaWallet } from '@tangled3/solana-react';
 import { useMutation } from '@tanstack/react-query';
 import { useCallback } from 'react';
 import { useConnect as useWagmiConnect } from 'wagmi';
+import { createTonWalletInstance } from '../connectors/ton/connector.js';
 import { useWalletsStore } from '../store/Wallet.js';
 import { ChainType } from '../types/index.js';
 import { DefaultConnector, Wallet, WalletInstance } from '../types/wallet.js';
 import { useAlephContext } from './useAlephContext.js';
+import { useTonContext } from './useTonContext.js';
 import { useTronContext } from './useTronContext.js';
 import { useWallets } from './useWallets.js';
 
@@ -20,6 +22,7 @@ export const useConnect = () => {
   const { connect: connectTronWallet } = useTronContext();
   const { connect: connectAlephWallet } = useAlephContext();
   const { mutate: connectSuiWallet } = useSuiConnectWallet();
+  const { connect: connectTonWallet } = useTonContext();
 
   const connectedWallets = useWalletsStore((state) => state.connectedWalletsByChain);
   const setCurrentWallet = useWalletsStore((state) => state.setCurrentWallet);
@@ -53,6 +56,12 @@ export const useConnect = () => {
         await connectAlephWallet(walletInstance.name);
       } else if (params.chainType === 'sui') {
         connectSuiWallet({ wallet: walletInstance.connector as WalletInstance<'sui'> });
+      } else if (params.chainType === 'ton') {
+        const connectedTonWallet = await connectTonWallet(walletInstance.id);
+        if (walletInstance.id === 'ton-connect') {
+          const tonWalletInstance = createTonWalletInstance(connectedTonWallet, walletInstance);
+          return { walletInstance: tonWalletInstance, name: tonWalletInstance.name, id: tonWalletInstance.id };
+        }
       } else {
         const connector = walletInstance.connector as DefaultConnector;
         await connector.connect();
@@ -68,6 +77,7 @@ export const useConnect = () => {
       connectTronWallet,
       connectedWallets,
       wallets,
+      connectTonWallet,
     ],
   );