+++ packages/react/src/hooks/useConnectionOrConfig.ts
@@ -4,6 +4,7 @@ import { useMemo } from 'react';
 import { useConfig as useWagmiConfig } from 'wagmi';
 import { ConnectionOrConfig } from '../types/index.js';
 import { useAlephStore } from './useAlephStore.js';
+import { useTonStore } from './useTonStore.js';
 import { useTronStore } from './useTronStore.js';
 
 /**
@@ -15,6 +16,7 @@ export const useConnectionOrConfig = (): ConnectionOrConfig | undefined => {
   const tronWeb = useTronStore((state) => state.tronweb);
   const alephZeroApi = useAlephStore((state) => state.api);
   const suiClient = useSuiClient();
+  const tonClient = useTonStore((state) => state.tonClient);
 
   return useMemo(() => {
     if (!alephZeroApi) return undefined;
@@ -25,6 +27,7 @@ export const useConnectionOrConfig = (): ConnectionOrConfig | undefined => {
       tronWeb,
       alephZeroApi: alephZeroApi,
       suiClient: suiClient,
+      tonClient,
     };
-  }, [wagmiConfig, solanaConnection, tronWeb, alephZeroApi, suiClient]);
+  }, [wagmiConfig, solanaConnection, tronWeb, alephZeroApi, suiClient, tonClient]);
 };