+++ packages/react/src/hooks/useDisconnect.ts
@@ -6,6 +6,7 @@ import { useDisconnect as useEVMDisconnect } from 'wagmi';
 import { ChainType } from '../types/index.js';
 import { DefaultConnector, Wallet, WalletInstance } from '../types/wallet.js';
 import { useAlephContext } from './useAlephContext.js';
+import { useTonContext } from './useTonContext.js';
 import { useTronContext } from './useTronContext.js';
 import { useWallets } from './useWallets.js';
 
@@ -20,6 +21,7 @@ export const useDisconnect = () => {
   const { disconnect: disconnectTronWallet } = useTronContext();
   const { disconnect: disconnectAlephWallet } = useAlephContext();
   const { mutate: disconnectSuiWallet } = useSuiDisconnectWallet();
+  const { disconnect: disconnectTonWallet } = useTonContext();
 
   const disconnectWallet = useCallback(
     async (params: DisconnectParams) => {
@@ -45,12 +47,14 @@ export const useDisconnect = () => {
         await disconnectAlephWallet();
       } else if (params.chainType === 'sui') {
         disconnectSuiWallet();
+      } else if (params.chainType === 'ton') {
+        await disconnectTonWallet();
       } else {
         const connector = walletInstance.connector as DefaultConnector;
         await connector.disconnect();
       }
     },
-    [disconnectAlephWallet, disconnectEVM, disconnectSolanaWallet, disconnectSuiWallet, disconnectTronWallet, wallets],
+    [disconnectAlephWallet, disconnectEVM, disconnectSolanaWallet, disconnectSuiWallet, disconnectTronWallet, disconnectTonWallet, wallets],
   );
 
   const mutation = useMutation({