+++ packages/react/src/providers/TangledContext.tsx
@@ -1,3 +1,4 @@
+import { TonConnectUIProvider } from '@tonconnect/ui-react';
 import { ReactNode, createContext, useState } from 'react';
 import { ChainData, ChainId, SupportedChainsByType, TangledConfig } from '../types/index.js';
 import { ChainConnectors } from '../types/wallet.js';
@@ -7,6 +8,7 @@ import { AlephProvider } from './AlephProvider.js';
 import EVMProvider from './EVMProvider.js';
 import { SolanaProvider } from './SolanaProvider.js';
 import { SuiProvider } from './SuiProvider.js';
+import { TonProvider } from './TonProvider.js';
 import { TronProvider } from './TronProvider.js';
 import WalletsProvider from './WalletsProvider.js';
 
@@ -36,6 +38,12 @@ export const TangledContextProvider = ({ children, config }: { children: ReactNo
   const [connectors] = useState(() => {
     return createChainConnectors(config, chains);
   });
+  const [tonconnectManifestUrl] = useState(() => {
+    return config.tonconnectManifestUrl;
+  });
+  const [twaReturnUrl] = useState(() => {
+    return config.twaReturnUrl;
+  });
 
   return (
     <TangledContext.Provider value={{ config, chains, connectors, chainsById }}>
@@ -50,7 +58,15 @@ export const TangledContextProvider = ({ children, config }: { children: ReactNo
           <SolanaProvider chain={chains.solana[0]}>
             <AlephProvider chain={chains.alephZero[0]}>
               <SuiProvider chains={chains.sui}>
-                <WalletsProvider>{children}</WalletsProvider>
+                {/* getting error if combined with TonProvider */}
+                <TonConnectUIProvider
+                  manifestUrl={tonconnectManifestUrl}
+                  actionsConfiguration={{ twaReturnUrl }}
+                >
+                  <TonProvider chain={chains.ton[0]}>
+                    <WalletsProvider>{children}</WalletsProvider>
+                  </TonProvider>
+                </TonConnectUIProvider>
               </SuiProvider>
             </AlephProvider>
           </SolanaProvider>