+++ packages/react/src/types/index.ts
@@ -1,12 +1,14 @@
 import { SuiClient, SuiTransactionBlockResponse } from '@mysten/sui/client';
 import { type ApiPromise } from '@polkadot/api';
 import { Connection as SolanaConnection } from '@solana/web3.js';
+import { TonClient } from '@ton/ton';
 import { GetTransactionReceiptReturnType as EVMTxReceipt } from '@wagmi/core';
 import { Types as TronWebTypes, type TronWeb } from 'tronweb';
 import { Chain as ViemChain } from 'viem';
 import { Config as WagmiConfig } from 'wagmi';
 import { CHAIN_ID } from '../constants/index.js';
 import { AlephTransactionData } from './aleph.js';
+import { TonTransactionInfo } from './ton.js';
 import { ChainConnectors } from './wallet.js';
 
 export const CHAIN_TYPES = [
@@ -19,6 +21,7 @@ export const CHAIN_TYPES = [
   'casper',
   'alephZero',
   'bitcoin',
+  'ton',
 ] as const;
 
 export type ChainType = (typeof CHAIN_TYPES)[number];
@@ -93,6 +96,11 @@ export interface TangledConfig {
   projectId: string;
 
   chainConnectors?: Partial<ChainConnectors>;
+
+  /** Manifest url for ton connect */
+  tonconnectManifestUrl?: string;
+  /** Telegram mini app url */
+  twaReturnUrl?: `${string}://${string}`;
 }
 
 type ChainRpcUrls = {
@@ -122,6 +130,7 @@ export type ConnectionOrConfig = {
   tronWeb: TronWeb;
   alephZeroApi: ApiPromise;
   suiClient: SuiClient;
+  tonClient: TonClient;
 };
 
 export type GetTokenMetadataParams = {
@@ -138,4 +147,6 @@ export type TransactionReceipt<C extends ChainType> = C extends 'evm'
       ? AlephTransactionData
       : C extends 'sui'
         ? SuiTransactionBlockResponse
-        : unknown;
+        : C extends 'ton'
+          ? TonTransactionInfo
+          : unknown;