+++ packages/react/src/types/wallet.ts
@@ -1,6 +1,7 @@
 import { WalletWithRequiredFeatures } from '@mysten/wallet-standard';
 import { NightlyConnectAdapter } from '@nightlylabs/wallet-selector-polkadot';
 import { Adapter as SolanaAdapter } from '@solana/wallet-adapter-base';
+import { TonConnectUI } from '@tonconnect/ui-react';
 import { Adapter as TronAdapter, AdapterState as TronAdapterReadyState } from '@tronweb3/tronwallet-abstract-adapter';
 import { Mutable } from '@wagmi/core/internal';
 import { CreateConnectorFn, Connector as EVMConnector } from 'wagmi';
@@ -22,6 +23,8 @@ export type ChainConnectors = {
   alephZero: any[];
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   bitcoin: any[];
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  ton: any[];
 };
 
 export type WalletBase<C extends ChainType = ChainType> = {
@@ -62,7 +65,9 @@ export type WalletInstance<T extends ChainType = ChainType> = T extends 'evm'
         ? NightlyConnectAdapter
         : T extends 'sui'
           ? WalletWithRequiredFeatures
-          : DefaultConnector;
+          : T extends 'ton'
+            ? TonConnectUI
+            : DefaultConnector;
 
 export type ConnectedWallet<T extends ChainType = ChainType> = {
   address: string;