+++ packages/react/src/utils/getDefaultSupportedChains.ts
@@ -2,6 +2,7 @@ import { alephZero } from '../chains/alephZero.js';
 import * as evm from '../chains/evm.js';
 import { solana } from '../chains/solana.js';
 import { sui } from '../chains/sui.js';
+import { tonMainnet } from '../chains/ton.js';
 import { tronMainnet } from '../chains/tron.js';
 import { EVMChain, OtherChainData, SuiChainType, SupportedChainsByType, TronChain } from '../types/index.js';
 
@@ -16,6 +17,7 @@ const getDefaultSupportedChains = (testnet?: boolean): SupportedChainsByType => 
     solana: [],
     sui: [],
     tron: [],
+    ton: [],
   };
 
   if (testnet) {
@@ -45,6 +47,7 @@ const getDefaultSupportedChains = (testnet?: boolean): SupportedChainsByType => 
     supportedChains.tron = [tronMainnet] as TronChain[];
     supportedChains.alephZero = [alephZero] as OtherChainData<'alephZero'>[];
     supportedChains.sui = [sui] as SuiChainType[];
+    supportedChains.ton = [tonMainnet] as OtherChainData<'ton'>[];
   }
 
   return supportedChains;